extracted out the common code from the AIA::Mods and AIA::Sgpt classes into a single module.
