+++ .semver
@@ -1,6 +1,6 @@
 ---
 :major: 0
 :minor: 3
-:patch: 22
+:patch: 23
 :special: ''
 :metadata: ''
