+++ lib/aia/tools/sgpt.rb
@@ -1,6 +1,9 @@
 # lib/aia/tools/sgpt.rb
 
+require_relative 'backend_common'
+
 class AIA::Sgpt < AIA::Tools
+  include AIA::BackendCommon
 
   meta(
     name:     'sgpt',
@@ -27,90 +30,6 @@ class AIA::Sgpt < AIA::Tools
     api_key
     # Add more directives if needed
   ]
-
-  attr_accessor :command, :text, :files, :parameters
-
-  def initialize(text: "", files: [])
-    @text       = text
-    @files      = files
-    @parameters = DEFAULT_PARAMETERS.dup
-    build_command
-  end
-
-
-  def sanitize(input)
-    Shellwords.escape(input)
-  end
-
-
-  def build_command
-    @parameters += " --model #{AIA.config.model} " if AIA.config.model
-    @parameters += AIA.config.extra
-
-    set_parameter_from_directives
-
-    @command = "sgpt #{@parameters} "
-    @command += sanitize(text)
-
-    puts @command if AIA.config.debug?
-
-    @command
-  end
-
-
-  # Clean up the temporary file after use
-  def clean_up_temp_file
-    @temp_file.unlink if @temp_file
-  end
-
-
-  def set_parameter_from_directives
-    AIA.config.directives.each do |directive, value|
-      if DIRECTIVES.include?(directive)
-        @parameters += " --#{directive} #{sanitize(value)}" unless @parameters.include?(directive)
-      end
-    end
-  end
-
-
-  def run
-    case @files.size
-    when 0
-      @result = `#{build_command}`
-    when 1
-      @result = `#{build_command} < #{@files.first}`
-    else
-      create_temp_file_with_contexts
-      run_with_temp_file
-      clean_up_temp_file
-    end
-    
-    @result
-  end
-
-
-  # Run with the temporary file as STDIN
-  def run_with_temp_file
-    command = "#{build_command} < #{@temp_file.path}"
-    @result = `#{command}`
-  end
-  
-
-  # Create a temporary file that concatenates all contexts,
-  # to be used as STDIN for the 'mods' utility
-  def create_temp_file_with_contexts
-    @temp_file = Tempfile.new('mods-context')
-
-    @files.each do |file|
-      content = File.read(file)
-      @temp_file.write(content)
-      @temp_file.write("\n")
-    end
-
-    @temp_file.close
-  end
-  
-
 end
 
 __END__
